fix: rename user to register
