+++ packages/app/src/features/register/data/MajorList.ts
